[Refactor] #120 - MainView 분리 완료.
