+++ PennyPack/App/PennyPackApp.swift
@@ -5,8 +5,7 @@ struct PennyPackApp: App {
     
     var body: some Scene {
         WindowGroup {
-            ContentView()
-
+            MainView(mainViewModel: MainViewModel(shoppingManager: ShoppingManager(), listManager: ListManager()))
         }
     }
 }
